feat(service): expose dead_code config via CLI
T036: Add --check-exported, --check-ffi, --edge-types parameters to the
dead_code CLI command. The run_dead_code function now accepts config
parameters and returns a DeadCodeOutput struct for testability.

Also fixes 14 clippy errors (Rust 1.96.0 new lints) across 6 files:
- field_reassign_with_default in build_dead_code_config
- manual_pattern_char_comparison in cross_service.rs
- get_first (row.get(0) -> row.first()) in trace/context.rs
- unnecessary_cast in query/structured.rs
- dead_code warnings for unused helper functions

13/13 tests pass, clippy clean.

+++ src/query/structured.rs
@@ -351,7 +351,7 @@ impl<'a> SearchEngine<'a> {
                     continue;
                 };
                 let node_id = row.first().and_then(|v| v.as_str()).unwrap_or("").to_string();
-                let degree = degree_map.get(&node_id).copied().unwrap_or(0) as u32;
+                let degree = degree_map.get(&node_id).copied().unwrap_or(0);
                 // Apply degree filter.
                 if let Some((min, max)) = params.degree_filter {
                     if degree < min || degree > max {
@@ -656,6 +656,7 @@ fn rows_to_search_results(
 /// - Prefix match → 0.8
 /// - Substring match → 0.5
 /// - No query (e.g. `search_by_type`) → 1.0 (neutral)
+#[allow(dead_code)]
 fn relevance_score(name: &str, query: &str) -> f64 {
     relevance_score_with_reason(name, query).0
 }
